+++ src/test/python/utilities/test_load_file_postprocessor.py
@@ -1,0 +1,10 @@
+import unittest
+
+
+class LoadFilePostprocessor(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, False)  # add assertion here
+
+
+if __name__ == '__main__':
+    unittest.main()
